move api router to dockyard/router.go and update README
